# models/admin.go
package models

import (
	"github.com/google/uuid"
	"gorm.io/gorm"
)

type Admin struct {
	gorm.Model
	ID       uuid.UUID `gorm:"type:char(36);primaryKey" json:"id"`
	Email    string    `                                json:"email"`
	Password string    `gorm:"type:char(60)"            json:"password"`
}

type CreateAdminRequest struct {
	Email     string `json:"email"`
	Password  string `json:"password"`
	SecretKey string `json:"secret_key"`
}